provider/openstack: revamp DestroyVolumes
Update DestroyVolumes to wait for volumes to be
detached before destroying.

+++ provider/openstack/cinder.go
@@ -6,6 +6,7 @@ package openstack
 import (
 	"math"
 	"net/url"
+	"sync"
 	"time"
 
 	"github.com/juju/errors"
@@ -15,6 +16,7 @@ import (
 
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/environs/tags"
+	"github.com/juju/juju/instance"
 	"github.com/juju/juju/storage"
 )
 
@@ -23,6 +25,11 @@ const (
 	// autoAssignedMountPoint specifies the value to pass in when
 	// you'd like Cinder to automatically assign a mount point.
 	autoAssignedMountPoint = ""
+
+	volumeStatusAvailable = "available"
+	volumeStatusDeleting  = "deleting"
+	volumeStatusError     = "error"
+	volumeStatusInUse     = "in-use"
 )
 
 type cinderProvider struct {
@@ -227,13 +234,63 @@ func (s *cinderVolumeSource) DescribeVolumes(volumeIds []string) ([]storage.Volu
 
 // DestroyVolumes implements storage.VolumeSource.
 func (s *cinderVolumeSource) DestroyVolumes(volumeIds []string) []error {
-	errors := make([]error, len(volumeIds))
+	var wg sync.WaitGroup
+	wg.Add(len(volumeIds))
+	results := make([]error, len(volumeIds))
 	for i, volumeId := range volumeIds {
-		if err := s.storageAdapter.DeleteVolume(volumeId); err != nil {
-			errors[i] = err
-		}
+		go func(i int, volumeId string) {
+			defer wg.Done()
+			results[i] = s.destroyVolume(volumeId)
+		}(i, volumeId)
 	}
-	return errors
+	wg.Wait()
+	return results
+}
+
+func (s *cinderVolumeSource) destroyVolume(volumeId string) error {
+	logger.Debugf("destroying volume %q", volumeId)
+	// Volumes must not be in-use when destroying. A volume may
+	// still be in-use when the instance it is attached to is
+	// in the process of being terminated.
+	var issuedDetach bool
+	volume, err := s.waitVolume(volumeId, func(v *cinder.Volume) (bool, error) {
+		switch v.Status {
+		default:
+			// Not ready for deletion; keep waiting.
+			return false, nil
+		case volumeStatusAvailable, volumeStatusDeleting, volumeStatusError:
+			return true, nil
+		case volumeStatusInUse:
+			// Detach below.
+			break
+		}
+		// Volume is still attached, so detach it.
+		if !issuedDetach {
+			args := make([]storage.VolumeAttachmentParams, len(v.Attachments))
+			for i, a := range v.Attachments {
+				args[i].VolumeId = volumeId
+				args[i].InstanceId = instance.Id(a.ServerId)
+			}
+			if len(args) > 0 {
+				if err := s.DetachVolumes(args); err != nil {
+					return false, errors.Trace(err)
+				}
+			}
+			issuedDetach = true
+		}
+		return false, nil
+	})
+	if err != nil {
+		return errors.Trace(err)
+	}
+	if volume.Status == volumeStatusDeleting {
+		// Already being deleted, nothing to do.
+		return nil
+	}
+	if err := s.storageAdapter.DeleteVolume(volumeId); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
 }
 
 // ValidateVolumeParams implements storage.VolumeSource.
@@ -312,7 +369,7 @@ func (s *cinderVolumeSource) DetachVolumes(args []storage.VolumeAttachmentParams
 		// Check to see if the volume is already detached.
 		attachments, err := s.storageAdapter.ListVolumeAttachments(string(arg.InstanceId))
 		if err != nil {
-			return err
+			return errors.Annotate(err, "listing volume attachments")
 		}
 		if err := detachVolume(
 			string(arg.InstanceId),
@@ -320,7 +377,10 @@ func (s *cinderVolumeSource) DetachVolumes(args []storage.VolumeAttachmentParams
 			attachments,
 			s.storageAdapter,
 		); err != nil {
-			return err
+			return errors.Annotatef(
+				err, "detaching volume %s from server %s",
+				arg.VolumeId, arg.InstanceId,
+			)
 		}
 	}
 	return nil
@@ -376,7 +436,10 @@ func newOpenstackStorageAdapter(environConfig *config.Config) (openstackStorage,
 		return nil, errors.Trace(err)
 	}
 
-	endpoint := authClient.EndpointsForRegion(ecfg.region())["volume"]
+	endpoint, ok := authClient.EndpointsForRegion(ecfg.region())["volume"]
+	if !ok {
+		return nil, errors.Errorf("volume endpoint not found for %q endpoint", ecfg.region())
+	}
 	endpointUrl, err := url.Parse(endpoint)
 	if err != nil {
 		return nil, errors.Annotate(err, "error parsing endpoint")
